The following changes are made:
1. Removed the "Local Flag: " prefix for flag descriptions
2. Added [x] prefix for all descriptions with the following meanings:
  g = global flag
  l = local flag, package not installed
  L = local flag, package is installed
  m = masked flag (package not installed if local)
  M = masked flag package is installed
3. Moved affected packages list of local flags to the front.

This is a preparation to add local/global and installed/uninstalled
filters.

The display, as it now is, is not pretty. I am not happy with the
static description line.

It is important to add a destinction between a globally masked flag,
and a locally masked flag. Further the prefix [x] is a) ugly and
b) must not scroll. It should be moved to the static left side,
getting its own column.